release.sh: generate the configure script/use unique build_dir
Allows us to simplify the releasing process of every component and
ensure that we're not rebuilding in a location which might contain stale
files.

As we can no longer pass arguments to configure (which in itself
is a bad idea when using `make distcheck') we gain more reproducible
releasing process.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>
---
XXX:
 - The whole "./build.XXX" and "../" feel quite nasty. We have to use
relative paths, since they end up in the generated sources. I'm open to
suggestions :-)

 - Should we cleanup the tmp/build dir ourselves, how about the
autotool generated files ?

 - If people prefer (or if things are broken due to required flags to
configure) we can guard this, or the original behaviour, behind a --dist
like option.